SavePoint_001

+++ src/repository.py
@@ -13,22 +13,31 @@ class Repository:
 
     def __init__(self):
         self.session = Session()
+        self.tables_d = {
+            User.__tablename__: User,
+            Administrator.__tablename__: Administrator,
+            Customer.__tablename__: Customer,
+            Flight.__tablename__: Flight,
+            Ticket.__tablename__: Ticket,
+            AirlineCompany.__tablename__: AirlineCompany,
+            Country.__tablename__: Country
+        }
 
-    def get_by_id(self, kind: object, _id: int) -> object:
+    def get_by_id(self, table_name: str, _id: int) -> object:
         """Finds the id given ,in a table."""
-        try:
-            obj = self.session.query(kind).get(_id)
+        table_name = '_'.join([w.capitalize() for w in table_name.split("_")])
+        if table_name in self.tables_d.keys():
+            table = self.tables_d[table_name]
+            obj = self.session.query(table).get(_id)
             if obj:
-                log.info(f"get_by_id() -> Finds object in {kind.__name__} table")
+                log.info(f"get_by_id() -> Finds object in {table_name} table")
                 return obj
             else:
-                log.info(f"get_by_id() -> cant find id: {_id} in {kind.__name__} table")
-                # print(f"ID: {_id} does not exist in {kind.__name__} table ")
-        except Exception as e:
-            log.info(f"get_by_id() -> Failed, make sure: {kind} object is an existing table.")
-            log.error(f"get_by_id() -> {e}")
+                log.info(f"get_by_id() -> cant find id: {_id} in {table_name} table")
+        else:
+            log.info(f"get_by_id() -> Failed, make sure: {table_name} is an existing table.")
 
-    def get_all(self, kind: object) -> list:
+    def get_all(self, kind: object) -> list:  ##TODO fix kind to table name
         """get all rows of a selected table into a list of obj"""
         try:
             table = self.session.query(kind).all()
@@ -38,7 +47,7 @@ class Repository:
             log.info(f"get_all() -> Failed and didn't find the table: {kind}")
             log.error(f"get_all() -> {e}")
 
-    def add(self, kind: object):
+    def add(self, kind: object):  ##TODO fix kind to table name
         """Add an object to a selected table"""
         try:
             self.session.add(kind)
@@ -48,11 +57,27 @@ class Repository:
             log.info(f"add() -> failed to add: {kind}")
             log.error(f"add() -> {e}")
 
-    def update(self, new_obj, kind, _id):  ## TODO after add
-        obj = self.get_by_id(kind, _id)
-        self.add()
+    def update(self, col: str, obj: object, val: str):  ## TODO after add
+        if type(obj) not in self.tables_d.values():
+            # print(type(obj))
+            # print(self.tables_d["Users"])
+            log.info(f"update() -> Make sure ur parameters are as it should be")
+            return None
+
+        attr = list(obj.__dict__.keys())[1:]
+        if col in attr:
+            try:
+                setattr(obj, col, val)
+                self.session.commit()
+                log.info(f"update() -> updated {obj.__tablename__} in column {col} id:{obj._id} ")
+            except Exception as e:
+                log.info(f"update() -> failed to update to: {val} \n\
+                make sure you know the constraints of this table")
+                log.error(f"update() -> {e}")
+        else:
+            log.info(f"update() -> Failed make sure that the column exist in table")
+
         # no use for update because it's not a real database for now
-        pass
 
     def add_all(self, object_list: object):
         """ Uses add() func on a list of objects."""
@@ -71,17 +96,25 @@ class Repository:
 
     def remove(self, obj: object):
         """Removes an object from our dictionary"""
-        try:
-            self.session.delete(obj)
-            log.info("remove() -> object deleted")
-            self.session.commit()
-        except Exception as e:
-            log.info(f"remove() -> this object doesn't exist in database \n{e}")
-            log.error(f"remove() -> {e}")
+        if obj:
+            try:
+                self.session.delete(obj)
+                log.info("remove() -> object deleted")
+                self.session.commit()
+            except Exception as e:
+                log.info(f"remove() -> Failed this object doesn't exist in database \n{e}")
+                log.error(f"remove() -> {e}")
+        else:
+            log.info(f"remove() -> Failed 'obj' Must be a table object ")
 
-
-    def get_airlines_by_country(self, country_id):
-        pass
+    def get_airlines_by_country(self, country: str) -> list:
+        airlines = self.session.query(AirlineCompany).\
+                filter(AirlineCompany.country.ilike(f"%{country}%"))
+        if airlines:
+            log.info(f"get_airlines_by_country() -> Found {len(list(airlines))} airlines in {country}")
+            return airlines
+        else:
+            log.info(f"get_airlines_by_country() -> Failed this 'country' parm doesnt exist in the database")
 
     def get_flights_by_origin_country_id(self, country_id):
         pass
@@ -114,7 +147,7 @@ class Repository:
 
 if __name__ == '__main__':
     repo = Repository()
-
+    repo.get_airlines_by_country("sdadsa22")
     # user_list = []
     # for i in range(10, 15):
     #     user_list += [User('Inbal_' + str(rnd(0, 99)), str(rnd(111111, 999999)), 'ILOVELeg' + str(i) + '@gmail.com', 1)]
